+++ esp_data/dataset/hf.py
@@ -1,3 +1,4 @@
+import logging
 import os
 from typing import Any, Callable, Generator, Iterable, Literal, Optional
 
@@ -5,13 +6,12 @@ from datasets import Dataset, concatenate_datasets, load_dataset, load_from_disk
 
 from esp_data.config.db_config import DataSample, DatasetConfig
 from esp_data.paths import AnyPath, make_storage_options
-from esp_data.utils import make_simple_logger
 
 from .base import BaseMapDataset
 from .shard_creator import write_huggingface_shard
 from .utils import generate_random_indices
 
-logger = make_simple_logger("esp_data")
+logger = logging.getLogger("esp_data")
 
 
 HF_DATASET_TYPES = [
