+++ esp_data/dataset/shard_creator.py
@@ -4,6 +4,7 @@ Modular functions to create sharded datasets in both WebDataset (tar) and Arrow 
 
 import hashlib
 import json
+import logging
 import time
 import traceback
 from concurrent.futures import ProcessPoolExecutor
@@ -24,11 +25,11 @@ import esp_data.file_io.functional as F
 from esp_data.config import DatasetConfig
 from esp_data.config.project_config import default_shard_creator_cfg
 from esp_data.paths import AnyPath, make_storage_options
-from esp_data.utils import make_id, make_simple_logger
+from esp_data.utils import make_id
 
 from .utils import make_file_opener
 
-logger = make_simple_logger("esp_data")
+logger = logging.getLogger("esp_data")
 
 # Initialize colorama for cross-platform colored terminal output
 colorama.init()
